requirements.txt
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyGithub==1.59.0
openai==0.27.8
tqdm==4.65.0
GitPython==3.1.31
